+++ tests/unit/asyn/wallet/test_wallet_generation.py
@@ -3,8 +3,7 @@ from unittest import TestCase
 from xrpl.asyncio.wallet.wallet_generation import (
     _AMM_DEV_FAUCET_URL,
     _DEV_FAUCET_URL,
-    _HOOKS_V2_TEST_FAUCET_URL,
-    _NFT_DEV_FAUCET_URL,
+    _HOOKS_V3_TEST_FAUCET_URL,
     _TEST_FAUCET_URL,
     get_faucet_url,
 )
@@ -36,14 +35,6 @@ class TestWallet(TestCase):
         self.assertEqual(get_faucet_url(json_client_url), expected_faucet)
         self.assertEqual(get_faucet_url(ws_client_url), expected_faucet)
 
-    def test_get_faucet_wallet_nft_dev(self):
-        json_client_url = "https://xls20-sandbox.rippletest.net:51234"
-        ws_client_url = "ws://xls20-sandbox.rippletest.net:51233"
-        expected_faucet = _NFT_DEV_FAUCET_URL
-
-        self.assertEqual(get_faucet_url(json_client_url), expected_faucet)
-        self.assertEqual(get_faucet_url(ws_client_url), expected_faucet)
-
     def test_get_faucet_wallet_amm_dev(self):
         json_client_url = "https://amm.devnet.rippletest.net:51233"
         ws_client_url = "wss://amm.devnet.rippletest.net:51233"
@@ -52,10 +43,10 @@ class TestWallet(TestCase):
         self.assertEqual(get_faucet_url(json_client_url), expected_faucet)
         self.assertEqual(get_faucet_url(ws_client_url), expected_faucet)
 
-    def test_get_faucet_wallet_hooks_v2_test(self):
-        json_client_url = "https://hooks-testnet-v2.xrpl-labs.com"
-        ws_client_url = "wss://hooks-testnet-v2.xrpl-labs.com"
-        expected_faucet = _HOOKS_V2_TEST_FAUCET_URL
+    def test_get_faucet_wallet_hooks_v3_test(self):
+        json_client_url = "https://hooks-testnet-v3.xrpl-labs.com"
+        ws_client_url = "wss://hooks-testnet-v3.xrpl-labs.com"
+        expected_faucet = _HOOKS_V3_TEST_FAUCET_URL
 
         self.assertEqual(get_faucet_url(json_client_url), expected_faucet)
         self.assertEqual(get_faucet_url(ws_client_url), expected_faucet)