+++ xrpl/models/requests/amm_info.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Dict
 
-from xrpl.models.currencies.issue import Issue
+from xrpl.models.currencies import Currency
 from xrpl.models.requests.request import Request, RequestMethod
 from xrpl.models.required import REQUIRED
 from xrpl.models.utils import require_kwargs_on_init
@@ -19,12 +19,12 @@ class AMMInfo(Request):
     Must provide Asset and Asset2 params.
     """
 
-    asset: Issue = REQUIRED  # type: ignore
+    asset: Currency = REQUIRED  # type: ignore
     """
     Specifies one of the pool assets (XRP or token) of the AMM instance.
     """
 
-    asset2: Issue = REQUIRED  # type: ignore
+    asset2: Currency = REQUIRED  # type: ignore
     """
     Specifies the other pool asset of the AMM instance.
     """