+++ xrpl/models/transactions/amm_bid.py
@@ -8,7 +8,7 @@ from typing_extensions import Final
 
 from xrpl.models.amounts import Amount
 from xrpl.models.auth_account import AuthAccount
-from xrpl.models.currencies.issue import Issue
+from xrpl.models.currencies import Currency
 from xrpl.models.required import REQUIRED
 from xrpl.models.transactions.transaction import Transaction
 from xrpl.models.transactions.types import TransactionType
@@ -28,12 +28,12 @@ class AMMBid(Transaction):
     discounted TradingFee for a 24 hour slot.
     """
 
-    asset: Issue = REQUIRED  # type: ignore
+    asset: Currency = REQUIRED  # type: ignore
     """
     Specifies one of the pool assets (XRP or token) of the AMM instance.
     """
 
-    asset2: Issue = REQUIRED  # type: ignore
+    asset2: Currency = REQUIRED  # type: ignore
     """
     Specifies the other pool asset of the AMM instance.
     """