+++ xrpl/models/transactions/xchain_create_bridge.py
@@ -16,13 +16,37 @@ from xrpl.models.xchain_bridge import XChainBridge
 @require_kwargs_on_init
 @dataclass(frozen=True)
 class XChainCreateBridge(Transaction):
-    """Represents a XChainCreateBridge transaction."""
+    """
+    Represents a XChainCreateBridge transaction.
+    The XChainCreateBridge transaction creates a new `Bridge` ledger object and
+    defines a new cross-chain bridge entrance on the chain that the transaction
+    is submitted on. It includes information about door accounts and assets for
+    the bridge.
+    """
 
     xchain_bridge: XChainBridge = REQUIRED  # type: ignore
+    """
+    The bridge (door accounts and assets) to create. This field is required.
+
+    :meta hide-value:
+    """
 
     signature_reward: str = REQUIRED  # type: ignore
+    """
+    The total amount to pay the witness servers for their signatures. This
+    amount will be split among the signers. This field is required.
+
+    :meta hide-value:
+    """
 
     min_account_create_amount: Optional[str] = None
+    """
+    The minimum amount, in XRP, required for a ``XChainAccountCreateCommit``
+    transaction. If this isn't present, the ``XChainAccountCreateCommit``
+    transaction will fail. This field can only be present on XRP-XRP bridges.
+
+    :meta hide-value:
+    """
 
     transaction_type: TransactionType = field(
         default=TransactionType.XCHAIN_CREATE_BRIDGE,
@@ -57,7 +81,7 @@ class XChainCreateBridge(Transaction):
                 "min_account_create_amount"
             ] = "Cannot have MinAccountCreateAmount if bridge is IOU-IOU."
 
-        if self.signature_reward is not None and not self.signature_reward.isnumeric():
+        if self.signature_reward != REQUIRED and not self.signature_reward.isnumeric():
             errors["signature_reward"] = "signature_reward must be numeric."
 
         if (