+++ xrpl/models/transactions/xchain_modify_bridge.py
@@ -3,9 +3,11 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
+from enum import Enum
 from typing import Dict, Optional
 
 from xrpl.models.currencies import XRP
+from xrpl.models.flags import FlagInterface
 from xrpl.models.required import REQUIRED
 from xrpl.models.transactions.transaction import Transaction
 from xrpl.models.transactions.types import TransactionType
@@ -13,16 +15,56 @@ from xrpl.models.utils import require_kwargs_on_init
 from xrpl.models.xchain_bridge import XChainBridge
 
 
+class XChainModifyBridgeFlag(int, Enum):
+    """
+    Transactions of the XChainModifyBridge type support additional values in the Flags
+    field. This enum represents those options.
+    """
+
+    TF_CLEAR_ACCOUNT_CREATE_AMOUNT = 0x00010000
+
+
+class XChainModifyBridgeFlagInterface(FlagInterface):
+    """
+    Transactions of the XChainModifyBridge type support additional values in the Flags
+    field. This TypedDict represents those options.
+    """
+
+    TF_CLEAR_ACCOUNT_CREATE_AMOUNT: bool
+
+
 @require_kwargs_on_init
 @dataclass(frozen=True)
 class XChainModifyBridge(Transaction):
-    """Represents a XChainModifyBridge transaction."""
+    """
+    Represents a XChainModifyBridge transaction.
+    The XChainModifyBridge transaction allows bridge managers to modify the
+    parameters of the bridge.
+    """
 
     xchain_bridge: XChainBridge = REQUIRED  # type: ignore
+    """
+    The bridge to modify. This field is required.
+
+    :meta hide-value:
+    """
 
     signature_reward: Optional[str] = None
+    """
+    The signature reward split between the witnesses for submitting
+    attestations.
+
+    :meta hide-value:
+    """
 
     min_account_create_amount: Optional[str] = None
+    """
+    The minimum amount, in XRP, required for a ``XChainAccountCreateCommit``
+    transaction. If this is not present, the ``XChainAccountCreateCommit``
+    transaction will fail. This field can only be present on XRP-XRP bridges.
+
+    :meta hide-value:
+    """
 
     transaction_type: TransactionType = field(
         default=TransactionType.XCHAIN_MODIFY_BRIDGE,